requirements.txt
ddddocr==1.4.3
Flask==2.0.3
gevent==21.12.0
gunicorn==20.1.0
PyExecJS==1.5.1
requests==2.27.1
